Server-Side Drive Speed
This is the next part in moving the state selection of the drive to the drive (server) side. The main benefit of this is in the future other external devices can easily send requests to update the state to the drive, and those can be handled efficiently. It also really condenses the remote code to make the remote a dumb terminal which only sends button info and displays values. More like a react paradigm.

BB8_remote.ino
  - Updated SpeedToggle to inherit from uint8_t (wire size)
  - Removed unneeded LastSpeedEntry
  - Renamed but6Speed to but6, as now it will simply send the raw value
  - Added bodySpeed to recFromBody. Note: BodySpeed is the SpeedToggle, DriveSpeed is the actual numeric value motor drive percentage, over 255. Remote only knows the former.
  - Removed setDriveSpeed() method from remote; it is moved to the drive now.
  - Display now toggles off recFromBody.bodySpeed
  - Display update check if lastDriveSpeed is different
  - Remove logic that does button click increment; simply send the raw button value
  - Not related, but renamed state to btConnectedState as I was getting it confused when coding

Main_Drive.ino
  - Add bodySpeed to sendToRemote
  - SpeedToggle is now uint8_t to match transfer variable
  - Add new button handler for button6, as the server is going to track the pushed state
  - remove but6 state vars
  - setDriveSpeed now calls incrementBodySpeedToggle(), which is the old setDriveSpeed method from the remote
  - setDriveSpeed now uses sendToRemote.bodySpeed instead of reading button6