Update Direction (Fwd/Reverse) to be driven from the Drive (Server)
This change drives the swap of forward and reverse to the server.

NOTE: The actual swapping of values is still performed on the client; in the future I could consider having the drive actually do the value swap as well, to have the remote as light as possible. But I feel this is already light enough for now, and is less chnage.

BB8_remote.ino
  - Add Direction enum to define forward and reverse
  - Rename Unknown for SpeedToggle to UnknownSpeed, and Direction has UnknownDirection. Prevents some compiler issues.
  - add bodyDirection to recFromBody
  - remove but5 state variables
  - remove reverseDirection; similar method will be implemented on the Drive (server)

  printVoltage()
    - read from recFromBody.bodyDirection when displaying the "Fwd/Rev" text.

  checkForScreenUpdate()
    - flag for update when the direction has changed from the last time

  readInputs()
    - directly send but5's value
    - key off recFromBody.bodyDirection to swap the remote values