Push To Roll Feature (AKA Safe Mode)
This feature is a new drive speed that only activates the drive motor for forward/reverse stability. What this means is that in this mode you can push the ball to have it roll to the desired ball orientation. This is useful for my case where there's really only one good circular panel that I can access; this allows pushing the ball to that one up without risking side to side or the head tilt causing an issue in the event of a catastrophic event for those axis.