Adjusting max angles
This change adjusts the max angles of the head tilt and dome tilt. I've been having some issues recently when driving with the head tilt getting locked in the forward position, but it seems to only be happening when I've been turning so I'm not sure which one is going out of bounds. So bringing both down for testing.