Safe Joystick Button Toggle
One major issue I was having was that when I was driving around I would accidently toggle reverse mode. This is pretty bad, as the drive would do some stuff you are not expecting, and if near objects can be catastrophic.

This fix ignores button presses when the joystick is being used. This prevents both the reverse issue above, and also from toggling dome server mode when using the dome.