Tuning based on continued testing of the drive.
When driving I noticed that the side to side was still a bit shakey, and that the head was too slow and not enough movement compared to my original tune. Since the new feature for Safe Joystick Button Toggle seemed to fix the issues I was having, I'm slightly increasing this to make sure it works as expected.