Feature: Head Tilt Stabilization
With this change the head tilt will now stay centered on top of the drive (mostly). This prevents the head from bobbing back before the drive starts moving which is one of the most non-BB-8 looking things about Joe's drive (and most drives for that matter). This works both if you have the TiltDomeForwardWhenDriving or not.