Code Cleanup - Create a PWM motor driver class that handles sending PWM to the motors.
Every motor driver method has a very similar code at the end of it to actually send the PWM to drive the motors. This change creates a class and helper method that are utilized to make the motor code generic. This allows the constants to be defined up on top of the class, making it easier to see, and limits the amount of duplicated code, making it easier to bring on new motors (not that I'll ever do that).