Head Stabilization Tuning
Adjust the head stabilization's proportion; the previous value was too strong, and could cause an oscillation to occur where the body and head were rocking continuously.

I also determined that the TiltDomeForwardWhenDriving was interferring with the stabilization. That's because it is taken right from the output of the motor. This should be updated to be proportional to the stick input for drive. For now I've just turned it off to avoid that issue.

Finally, there is just some code cleanup for some #ifdef statements and one if has been converted to a trinary just to make the code a bit more clean.