Change proportion to be over 1.0
Originally I thought that maybe the proportion would need to be above 1.0, but that is not the case. Thus, I've changed the proportion to be out of a 1.0 scale to make it more readible, as now setting it to 0.8 means 80%. You can still do like 1.2 if you somehow need more.