Dome Modes (might need to rename)
I added the new mode toggling via the left thumbstick. This includes some modes that are not fully flushed out, but the two modes that have tilt do honor the tilt. Further, tilt is now based on the drive joystick position as opposed to the motor output. This prevents the bobble loop, and allows the head to move forward much quicker, and it has a well defined constant now.

I updated the remote as well to display all the new modes. I MIGHT need to remove the OFF variable, as it bounces to that. But it might be because I have my voltage divider disconnected and that could be causing some issues.