Create a Joystick Ease applicator class (card https://github.com/naigonakoii/JoesDrive/projects/1#card-40370389)
This change removes the manual ease applications and manual table Joe created for the drive curve, and replaces all the hard-coded if-else statements that applied the ease functionality throughout the code. It really cleans up all the motor driving code and allows the constants to just be at the object creation (besides the constrain functions).

There was also some dead code in the flywheel specifically that I removed.

Finally, I also converted all drive methods to utilize the #ifdef pattern for reverse. This removes the need for #ifdef blocks down in those methods as now it is taken care of by the revF1 revF2 pairs for each feature F.

TODO - I need to test a bit but since Joe's S-curve was pretty wide and slow didn't get to the end it was more like a quadratic. I think I'll need to allow a quadratic motion for slow and medium to better replicate what makes sense for drive.