Fix for Issue #8 - Cancel for Body/Dome calibration not working correctly
There were a couple parts to this change. First, I fixed the remote code, as there is no longer any need for the calibrationMarker variable.

Then, I changed the behaivor on the drive. Instead of having the counter which made the code a bit convoluted, I simply have right most button (button8) be accept and the button to the left of it (button7) be cancel, like B and A on XBox.

Also in this change is moving out the debug method into its own .ino for ease of reading, and a little bit more minor refactory to shrink the overall number of lines. I have a goal to get the main file under 1000 lines; right now it is at 1200. I may cheat and make a separate .ino file for all the motion functions since those are the meat of the code and it would be nice to have them in a separate tab.