Add a Joystick Wrapper - (https://github.com/naigonakoii/JoesDrive/projects/1#card-40861990)
This is a refactory to remove all the individual map(...) functions for the joystick and pot inputs, and to wrap them in a new class called AnalogInHandler. This class acts similar to the ButtonHandler, in that every loop iteration it is updated, then the mapped value can be pulled from there-on in the rest of the loop.

The AnalogInHandler takes care of all the input to output mapping, including reversing the order to the mapped output via passing a flag. This cleaned up the reverse constants a lot as now they are simple true/false definitions; there is no need for the rev1 and rev2 pairs.

The AnalogInHandler also now is able to flag if there was any perceived movement for that input, which simplifies the Safe Joystick and Joe's existing power off methods a bit.

One thing that confused me in Joe's code is that enabling reverse didn't affect writing into the EEPROM. I switched that and the limited testing I could do (swapping my dome spin) seemed to work correctly doing it this way. However, it is possible that there will be an issue with this approach that I am not currently thinking of.

Also, I did a proposed fix for Issue #7 as part of the refactory, as now in Safe Mode I no longer need to override the joysticks as part of the 'movement' method. Along with that, I updated disableAllTheThings to selectively disable the main motor so that others can be disabled in safe mode.