Initial Animations
This is a big and critical change to update the animations that was partially in the build to being a robust system.

Animations now work on a simple state mechanim. An animation is just an array of AnimationState, where AnimationState is a class (basically a Struct) that contains the metadata for how the joysticks should be operated, if a sound should be played, and how long to stay on the current state. The animation class is responsible for maintaining which AnimationState is active and returning that up. It also deals with auto stopping when there are no more states.

The AnimationManager maintains the collection of every Animation known to the system. It also deals with the two (currently two; could add more) unique animation types, and is responsible for selecting a random animation from the specified type and starting it to run. It knows when this current animation is done as well, and passes back the states along the way.

In the main program, I currently only have button 4 press hooked up to play any animation; I will eventually change this to only the full for button 4 press, and domes for button 4 hold (or maybe button 7 press, as that's open as well). I also need to hook up to randomly do the dome animations when in automated mode.

Finally, I hooked up a new feature at the top level to force the dome spin into server mode when an animation completes. This puts the head back on center after the animation runs, which is good for driving around.