Refactor - struct grouping
In an effort to clean up the code a bit more I moved all the dangling global variables (save for the PID ones which I will deal with later) into logical structs. This helped to really get a clear view on how the globals were being used, and made the following 3 main areas emerge:

IMU
Need to refactor an IMU wrapper class that the program can call a RunIteration on, and it will provide the updated values.

Offsets
Refactor these into a class that has a save method as well

Drive
All the methods Movement calls need to become a big class. Will need all PIDs passed in