Ez Toggle between Automated vs normal
This is a minor change that makes going into automation quicker. Since all three right buttons are free at this point, I decided to use one of them to switch between automated vs not. When toggling this, then the left stick will advance through the automated modes if auto is enabled, or manual modes if it isn't. This makes it very quick to get between a normal and automated mode, that way if I'm driving and tons of people start hording I can turn off auto and control manually.