Animations - Animation Step supports setting the dome spin mode
Some animations look better with the dome spin in servo, but others look better with it in full spin. To not rely on the current drive mode, the AnimationStep has now been updated to indicate the desired AnimationDomeMode in addition to the original settings. This mode can be set to adSpin, adServo, or adEither; the latter will retain the mode the drive is currently using. GeneratedAnimation now takes a parameter of which to use, and for ScriptedAnimation instances, each entry in the passed in AnimationStep[] array can have the appropriate value set. Note that it can change each animation step but when a change occurs it is a bit jerky so I recommend minimizing those changes (ie keep the same through the entire animation or change at most once).

This is a really nice addition to animations that really rounds it out, and now I'm pleased overall, just want to add a few more scripted ones to button 4.