All AnimationRunner to run a specific id
All Animations when created will now get an id assigned. The IAnimation provides a getter to this id. This id can be passed back into the new AnimationRunner.RunAnimationWithId() method.

In the drive, this is now used to switch to the specific animation depending on whether the user is in automated spin or automated servo; the need for specific animations here came from testing and realizing that it isn't great to be driving along and have the head start spinning rapidly, so specific parameters need to be passed in to the GeneratedAnimationPercents for the LR stick to make it way tuned down for the full spin case. For the button press animations, this isn't a big issue because the droid should be sitting still when pressing the button.

Animation cpp/h
  - Add "uint16_t Id() const" to interface and implement it in both scripted and generated implementations.
  - Add "uin16_t GlobalId" in cpp file
  - Constructor for GeneratedAnimation and ScriptedAnimation both pass in GlobalId++ to get the unique id.

AnimationRunner h/cpp
  - add StartAnimationWithId method

Animations.ino
  - Create seperate animations for head spin and head servo in bank1 (automation)

Main_Drive.ino
  - Add logic that stops the running animation when the animated mode changes
  - Find the id and use the new RunAnimationWithId() method from AnimationRunner if an animation isn't running in Automation mode.