Fix animations causing head to pop
The issue I was having was related to PID tuning and easing. Even with a function ease applicator, the ease is static in that the end can still have a hard stop. Instead, I created an ease which gradually increased/decreses the increment like what the linear one does. This allows the value to slow down before it reaches its target, which makes way less issues with head pops.

Still needs a bit of tuning but the general code is in and working.