Fix PSP sticks to have full range
I upaded based on the average of my two sticks. This will be MUCH closer than no check as those PSP sticks don't have a full range on their potentiometers.

Note: I also confirmed that the wiring for the positive and negative rails was backwards on the two PSP sticks. Changing those fixed the leak issue of those two potentometers.