Fix side to side stability issues.
These are actually some bugs that are existing in the master branch as well, and should be ported there. These are what was causing the weird side to side stability issues in my BB-8. However, it was much easier to test with Joe's BB-9E, and as such I fixed the issues here.

Note: I did investigate making one PID loop as opposed to two, and I think it is actually possible using the subtraction of the rollDifference from the pot difference as the PID input. However, it seemed to react slower on the ground, though when holding it up it seemed to work. I'll leave that as an item to play with after MK3 gets merged into master.

ImuProMini.h/.cpp
  - Remove the filtering from this class; filtering should actually be done later in motor methods as it is desireable to feed the exact value into the PID loop. Otherwise it can drift over time.

Offsets.cpp
  - This was the main bug that was messing with S2S; I fixed setting the write count specifically for update values so they will always get written correctly.

Constants.h
  - Adjust ease for new PID
  - Add specific ease for MK3 since the dome is a completely different motor (servos) so is understandably different.
  - Readjust PID values for S2S now that it is fixed

Main_Drive.ino
  - Go back to a linear applicator for side to side. It makes sense to allow the PID control here for the setpoint
  - Wrap PID4 in #if HeadTiltVersion == MK2_Dome; it is completely not needed in this drive

Debug.ino
  - Fix the S2S and DomeSpin debug #ifdefs since I needed them.