MK3 specific Dome Tilt enhancements
Two major things are added with this change:
1. Tuning the MK3 dome so the ease is a bit better. NOTE: since ease is applied on X and Y independently, some angular motions are not eased as well.
2. Implementing the pitch and roll inputs to enable the auto-head leveling

There was also a bug where if the Y had not been moved for the head tilt, it would go way past the X end stops. To fix, I created a little bit of code that is similar but different than Joe's to ensure that the head moves in the Y direction if it is moved in the X direction.

Structs.h
  - Add HasDomeMoved property to the drive struct

Constants.h
  - Ease parameters for the X and Y dome tilt
  - Updated the roll proportion to .9 from .8

Main_Drive.ino
  - Update the two domeTiltEaseFR/LR vars with new constants
  - Initialize HasDomeMoved
  - Add adjust variables into the X and Y joy values
  - Logic to move the Y by 2 when moving the X if the Y has never been moved by the user.