Joe's Serial Audio port from the MK3 Head Tilt official release.
This is mostly a port of the code, but I changed a few things:
  - I removed the concept of quick files; I don't like the experience of tapping a button multiple times to give a specific sound; with the sounds being ordered it should be good enough.
  - Left button1 press is sounds in order; left button1 hold is random
  - Left button2 press is play the next music sound
  - Left button2 hold will stop any playing music sound

Also, I changed my code to use the sfx class wrappers instead of directly manipulating the Serial3 stream. This allows swapping serial2 and serial3 as older Joe's diagrams did have the separate IMU on serial 3.

Finally, I slightly changed the communication. Joe's standalone feather would read two separate pushes from the main drive, one that was for the dome and one that was for the remote. However, since I really only need to send the PSI to the dome, I've updated my code to just push the PSI along with the data that goes to the remote. Then, the standalone feather just breaks apart the data and sends along two separate calls, one to the remote and one to the dome. In order to still allow the Bluetooth remote to not have PSI I thus needed to create this separate instance, and then some #if RemoteHardware wrappers are used.

Structs.h
  - Create a new SEND_DATA_STRUCTURE_FEATHER. This one includes the psi value that will be send when using feathers.
  - Rename AudioParams to SerialAudioParams, and update to those needed only by the serial audio.

MK3_Feather_Standalone.ino
  - Create the receive data structure to match the new one the body sends (with PSI). The existing one stays as being sent to remote.
  - Update the code to copy the values from the received body to the sendToRemote and sendToDome structs.

Main_Drive.ino
  - Include adafruit's soundboard class.
  - #if RemoteHardware switch to use the correct send structure.
  - Setup for new audio including Serial3 (or 2) and the sfx class init.

  - The following is wrapped behind #if AudioHardware == JoeSerial
    - Add the psiVal method back which adds the PSI val to the global send.
    - Add a writeSerialSound(int num) method. This uses the sfx class to stop a sound if playing and play the requested one. (Joe's code would ignore new requests).
    - Add the main audio handling that plays sounds and music from left button1 and left button2 (button2 and button3).