Fix for Issue #14, take 2
This time I did a much simpler fix, and safe-guarded the body calibration for the motor enable to try and prevent it from kicking in and getting stuck. I also moved the buttons for joystick calibration to L button 3 and R button 1, the inner two most buttons to not conflict with the right most button which acts as selecting.