Remote fixes
I noticed that when the body was on low-power (only plugged with the USB) it was getting some pretty low voltages that flucuate quickly. To prevent this, I made voltage now something that only gets updated on the periodic refresh. I also changed it so that period is 8 seconds instead of 15, which I felt was a bit long.