FINALLY! Issue #25 resolved
The main culprut here was actually that if the serial speed for the oled screen was set to high its processing delays could overtake the radio and basically deadlock the program. By reducing this down from 400,000L to 100,000L, the issue was MOSTLY resolved. I found that there were still a few minor timing bugs so I fixed those as well. The result seems to be stable communications with the remote and drive now.

MK3_Feather_Standalone.ino
  - Add DebugBody constant to help debugging (mostly for me)
  -Remove some unused variables
  -Ensure send was happening at a sustainable pace - was too frequent before.

MK3_Remote.ino
  -Update send to happen every 50ms
  -Read inputs right before send only to ensure most current values
  -Skip sending if the radio is blocked
  -Check for receive now happens every time (put back a change I backed out before)
  -Added DebugRecBody and some extra debugging